Added some functions working with collection
These functions are not implemented now.
Moved windows definitions to millandict.py.
We should reduce number of files a little.
Because Anki sees every file as a plugin
and it is strange when you see every file
of your plugin as a separate plugin.
We should stay with two files: a plugin file
and a MillanDict API wrapper file.